+++ docs/com/marginallyclever/communications/TransportLayerPanel.html
@@ -113,7 +113,7 @@ var activeTableTab = "activeTableTab";
 <li><a href="http://docs.oracle.com/javase/8/docs/api/javax/swing/JPanel.html?is-external=true" title="class or interface in javax.swing">javax.swing.JPanel</a></li>
 <li>
 <ul class="inheritance">
-<li>com.marginallyclever.communications.TransportLayerPanel</li>
+<li>com.marginallyclever.communications.transport.TransportLayerPanel</li>
 </ul>
 </li>
 </ul>
