+++ docs/com/marginallyclever/communications/serial/SerialTransportLayer.html
@@ -101,7 +101,7 @@ var activeTableTab = "activeTableTab";
 <li><a href="http://docs.oracle.com/javase/8/docs/api/java/lang/Object.html?is-external=true" title="class or interface in java.lang">java.lang.Object</a></li>
 <li>
 <ul class="inheritance">
-<li>com.marginallyclever.communications.serial.SerialTransportLayer</li>
+<li>com.marginallyclever.communications.transport.serial.SerialTransportLayer</li>
 </ul>
 </li>
 </ul>
