+++ docs/com/marginallyclever/communications/tcp/TCPConnection.html
@@ -101,7 +101,7 @@ var activeTableTab = "activeTableTab";
 <li><a href="http://docs.oracle.com/javase/8/docs/api/java/lang/Object.html?is-external=true" title="class or interface in java.lang">java.lang.Object</a></li>
 <li>
 <ul class="inheritance">
-<li>com.marginallyclever.communications.tcp.TCPSession</li>
+<li>com.marginallyclever.communications.transport.tcp.TCPSession</li>
 </ul>
 </li>
 </ul>
