+++ docs/com/marginallyclever/communications/tcp/TCPTransportLayerPanel.html
@@ -113,10 +113,10 @@ var activeTableTab = "activeTableTab";
 <li><a href="http://docs.oracle.com/javase/8/docs/api/javax/swing/JPanel.html?is-external=true" title="class or interface in javax.swing">javax.swing.JPanel</a></li>
 <li>
 <ul class="inheritance">
-<li><a href="../../../../com/marginallyclever/communications/TransportLayerPanel.html" title="class in com.marginallyclever.communications">com.marginallyclever.communications.TransportLayerPanel</a></li>
+<li><a href="../../../../com/marginallyclever/communications/TransportLayerPanel.html" title="class in com.marginallyclever.communications">com.marginallyclever.communications.transport.TransportLayerPanel</a></li>
 <li>
 <ul class="inheritance">
-<li>com.marginallyclever.communications.tcp.TCPTransportLayerPanel</li>
+<li>com.marginallyclever.communications.transport.tcp.TCPTransportLayerPanel</li>
 </ul>
 </li>
 </ul>
