+++ docs/com/marginallyclever/communications/tcp/class-use/TCPConnection.html
@@ -4,7 +4,7 @@
 <head>
 <!-- Generated by javadoc (1.8.0_66) on Mon Feb 24 12:02:36 PST 2020 -->
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>Uses of Class com.marginallyclever.communications.tcp.TCPSession (Robot Overlord 1.6.0 API)</title>
+<title>Uses of Class com.marginallyclever.communications.transport.tcp.TCPSession (Robot Overlord 1.6.0 API)</title>
 <meta name="date" content="2020-02-24">
 <link rel="stylesheet" type="text/css" href="../../../../../stylesheet.css" title="Style">
 <script type="text/javascript" src="../../../../../script.js"></script>
@@ -71,9 +71,9 @@
 </a></div>
 <!-- ========= END OF TOP NAVBAR ========= -->
 <div class="header">
-<h2 title="Uses of Class com.marginallyclever.communications.tcp.TCPSession" class="title">Uses of Class<br>com.marginallyclever.communications.tcp.TCPSession</h2>
+<h2 title="Uses of Class com.marginallyclever.communications.transport.tcp.TCPSession" class="title">Uses of Class<br>com.marginallyclever.communications.transport.tcp.TCPSession</h2>
 </div>
-<div class="classUseContainer">No usage of com.marginallyclever.communications.tcp.TCPSession</div>
+<div class="classUseContainer">No usage of com.marginallyclever.communications.transport.tcp.TCPSession</div>
 <!-- ======= START OF BOTTOM NAVBAR ====== -->
 <div class="bottomNav"><a name="navbar.bottom">
 <!--   -->
