+++ docs/serialized-form.html
@@ -81,7 +81,7 @@
 <li class="blockList"><a name="com.marginallyclever.communications.TransportLayerPanel">
 <!--   -->
 </a>
-<h3>Class <a href="com/marginallyclever/communications/TransportLayerPanel.html" title="class in com.marginallyclever.communications">com.marginallyclever.communications.TransportLayerPanel</a> extends <a href="http://docs.oracle.com/javase/8/docs/api/javax/swing/JPanel.html?is-external=true" title="class or interface in javax.swing">JPanel</a> implements Serializable</h3>
+<h3>Class <a href="com/marginallyclever/communications/TransportLayerPanel.html" title="class in com.marginallyclever.communications">com.marginallyclever.communications.transport.TransportLayerPanel</a> extends <a href="http://docs.oracle.com/javase/8/docs/api/javax/swing/JPanel.html?is-external=true" title="class or interface in javax.swing">JPanel</a> implements Serializable</h3>
 <dl class="nameValue">
 <dt>serialVersionUID:</dt>
 <dd>8909954928628121764L</dd>
@@ -95,7 +95,7 @@
 <li class="blockList"><a name="com.marginallyclever.communications.serial.SerialTransportLayerPanel">
 <!--   -->
 </a>
-<h3>Class <a href="com/marginallyclever/communications/serial/SerialTransportLayerPanel.html" title="class in com.marginallyclever.communications.serial">com.marginallyclever.communications.serial.SerialTransportLayerPanel</a> extends <a href="com/marginallyclever/communications/TransportLayerPanel.html" title="class in com.marginallyclever.communications">TransportLayerPanel</a> implements Serializable</h3>
+<h3>Class <a href="com/marginallyclever/communications/serial/SerialTransportLayerPanel.html" title="class in com.marginallyclever.communications.serial">com.marginallyclever.communications.transport.serial.SerialTransportLayerPanel</a> extends <a href="com/marginallyclever/communications/TransportLayerPanel.html" title="class in com.marginallyclever.communications">TransportLayerPanel</a> implements Serializable</h3>
 <dl class="nameValue">
 <dt>serialVersionUID:</dt>
 <dd>-5048852192781164326L</dd>
@@ -124,7 +124,7 @@
 <li class="blockList"><a name="com.marginallyclever.communications.tcp.TCPTransportLayerPanel">
 <!--   -->
 </a>
-<h3>Class <a href="com/marginallyclever/communications/tcp/TCPTransportLayerPanel.html" title="class in com.marginallyclever.communications.tcp">com.marginallyclever.communications.tcp.TCPTransportLayerPanel</a> extends <a href="com/marginallyclever/communications/TransportLayerPanel.html" title="class in com.marginallyclever.communications">TransportLayerPanel</a> implements Serializable</h3>
+<h3>Class <a href="com/marginallyclever/communications/tcp/TCPTransportLayerPanel.html" title="class in com.marginallyclever.communications.tcp">com.marginallyclever.communications.transport.tcp.TCPTransportLayerPanel</a> extends <a href="com/marginallyclever/communications/TransportLayerPanel.html" title="class in com.marginallyclever.communications">TransportLayerPanel</a> implements Serializable</h3>
 <dl class="nameValue">
 <dt>serialVersionUID:</dt>
 <dd>-5157947915933861665L</dd>
