+++ src/main/java/com/marginallyclever/communications/transport/serial/SerialTransportLayer.java
@@ -1,8 +1,8 @@
-package com.marginallyclever.communications.serial;
+package com.marginallyclever.communications.transport.serial;
 
 import com.marginallyclever.communications.session.SessionLayer;
-import com.marginallyclever.communications.TransportLayer;
-import com.marginallyclever.communications.TransportLayerPanel;
+import com.marginallyclever.communications.transport.TransportLayer;
+import com.marginallyclever.communications.transport.TransportLayerPanel;
 import jssc.SerialPortList;
